requirements.txt
aiogram==3.10.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-telegram-bot==20.7